IAT-480: Refactor HTTP error handling from services to shared utility.
