+++ src/app/core/services/app-info/app-info.service.ts
@@ -2,17 +2,21 @@ import {Injectable} from "@angular/core";
 import {Http} from "@angular/http";
 import {Observable} from "rxjs/Observable";
 import {AppInfo} from "./app-info";
+import {HttpUtility} from "../../util/http-utility";
+import {Logger} from "../logger/logger.service";
 
 @Injectable()
 export class AppInfoService {
   private serviceUrl = '/manage/info';
 
-  constructor(private http: Http) {
+  constructor(private http: Http,
+              private logger: Logger) {
   }
 
   findAppInfo(): Observable<AppInfo> {
     return this.http
       .request(this.serviceUrl)
-      .map(request => AppInfo.fromDto(request.json()));
+      .map(request => AppInfo.fromDto(request.json()))
+      .catch(error => HttpUtility.logAndThrowError(this.logger, error));
   }
 }
